requirements.txt
paho.mqtt
psycopg2
